requirements.txt
pandas
numpy
scikit-learn
xgboost
fastapi
uvicorn
pydantic
pyyaml
mlflow
dvc
evidently
pytest
python-dotenv
matplotlib